// src/apps/external-app/pendo/pendo.types.ts
interface IPendoOptions {
  orgCode: string;
  organizationName: string;
  customerType: string;
  userEmailAddress: string;
  userId: string;
  userRole: string;
  tenantIndustry: string;
  tenantSubIndustry: string;
  tenantBusinessType: string;
  accountPlan: string;
  regionId: string;
  clusterId: string;
  userCreationDate: string;
  isSystemUser: string;
  creationDate: string;
  tenantRenewalDate: string;
  tenantCity: string;
  tenantCountry: string;
  tenantSuccessOwner: string;
  tenantAccountOwner: string;
  isAgency: string;
  accountTimeZoneValue: string;
  tenantBillingOption: string;
  userRegionId: string;
  userClusterId: string;
  apiKey: string;
  loginAppUrl: string;
  userDesignation: string;
  appType: string;
  accountStartingDate: string;
}

export type { IPendoOptions };